This PR includes refactoring export function and update to azure storage blob v12.

Changes Included :
1)[Refactoring Export function 1448291](https://domoreexp.visualstudio.com/MSTeams/_workitems/edit/1448291)

1) ExportOrchestration - Removed the dependencies from the orchestrator and invoked the activity directly.
2) Activity classes - Refactored the activity classes to only have activity function and removed the earlier pattern of having another function using context to call the activity.
3) Refactored classes to remove concrete dependency and make them unit testable.

2)[Upload Azure storage blob v12 1465479](https://domoreexp.visualstudio.com/MSTeams/_workitems/edit/1465479)

Updated the Azure storage package from v11 to v12. As part of this rewritten the upload function in Upload Activity using v12 package. The upload strategy is same as before and no change to the functionality.

3)Updated unit test

Verification:
Pending for large data.

+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Prep.Func/Export/Activities/HandleExportFailureActivity.cs
@@ -8,17 +8,15 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Prep.Func.Export.Activities
     using System;
     using System.Threading;
     using System.Threading.Tasks;
-    using global::Azure.Storage.Blobs;
     using Microsoft.Azure.WebJobs;
     using Microsoft.Azure.WebJobs.Extensions.DurableTask;
     using Microsoft.Bot.Builder;
-    using Microsoft.Bot.Builder.Integration.AspNet.Core;
     using Microsoft.Bot.Connector.Authentication;
     using Microsoft.Bot.Schema;
     using Microsoft.Extensions.Localization;
-    using Microsoft.Extensions.Logging;
     using Microsoft.Extensions.Options;
-    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Adapter;
+    using Microsoft.Teams.Apps.CompanyCommunicator.Common.Clients;
     using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.ExportData;
     using Microsoft.Teams.Apps.CompanyCommunicator.Common.Repositories.UserData;
     using Microsoft.Teams.Apps.CompanyCommunicator.Common.Resources;
@@ -33,55 +31,35 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Prep.Func.Export.Activities
     public class HandleExportFailureActivity
     {
         private readonly IExportDataRepository exportDataRepository;
-        private readonly string storageConnectionString;
-        private readonly BlobContainerClient blobContainerClient;
+        private readonly IStorageClientFactory storageClientFactory;
         private readonly IUserDataRepository userDataRepository;
         private readonly string authorAppId;
-        private readonly BotFrameworkHttpAdapter botAdapter;
+        private readonly ICCBotFrameworkHttpAdapter botAdapter;
         private readonly IStringLocalizer<Strings> localizer;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="HandleExportFailureActivity"/> class.
         /// </summary>
         /// <param name="exportDataRepository">the export data respository.</param>
-        /// <param name="repositoryOptions">the repository options.</param>
+        /// <param name="storageClientFactory">the storage client factory.</param>
         /// <param name="botOptions">the bot options.</param>
         /// <param name="botAdapter">the users service.</param>
         /// <param name="userDataRepository">the user data repository.</param>
         /// <param name="localizer">Localization service.</param>
         public HandleExportFailureActivity(
             IExportDataRepository exportDataRepository,
-            IOptions<RepositoryOptions> repositoryOptions,
+            IStorageClientFactory storageClientFactory,
             IOptions<BotOptions> botOptions,
-            BotFrameworkHttpAdapter botAdapter,
+            ICCBotFrameworkHttpAdapter botAdapter,
             IUserDataRepository userDataRepository,
             IStringLocalizer<Strings> localizer)
         {
-            this.exportDataRepository = exportDataRepository;
-            this.storageConnectionString = repositoryOptions.Value.StorageAccountConnectionString;
-            this.blobContainerClient = new BlobContainerClient(this.storageConnectionString, Common.Constants.BlobContainerName);
-            this.botAdapter = botAdapter;
-            this.authorAppId = botOptions.Value.AuthorAppId;
-            this.userDataRepository = userDataRepository;
+            this.exportDataRepository = exportDataRepository ?? throw new ArgumentNullException(nameof(exportDataRepository));
+            this.storageClientFactory = storageClientFactory ?? throw new ArgumentNullException(nameof(storageClientFactory));
+            this.botAdapter = botAdapter ?? throw new ArgumentNullException(nameof(botAdapter));
+            this.authorAppId = botOptions?.Value?.AuthorAppId ?? throw new ArgumentNullException(nameof(botOptions));
+            this.userDataRepository = userDataRepository ?? throw new ArgumentNullException(nameof(userDataRepository));
             this.localizer = localizer ?? throw new ArgumentNullException(nameof(localizer));
-        }
-
-        /// <summary>
-        /// Run the activity.
-        /// </summary>
-        /// <param name="context">Durable orchestration context.</param>
-        /// <param name="exportDataEntity">export data entity.</param>
-        /// <param name="log">Logging service.</param>
-        /// <returns>instance of metadata.</returns>
-        public async Task RunAsync(
-            IDurableOrchestrationContext context,
-            ExportDataEntity exportDataEntity,
-            ILogger log)
-        {
-            await context.CallActivityWithRetryAsync<Task>(
-                      nameof(HandleExportFailureActivity.HandleFailureActivityAsync),
-                      FunctionSettings.DefaultRetryOptions,
-                      exportDataEntity);
         }
 
         /// <summary>
@@ -91,10 +69,15 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Prep.Func.Export.Activities
         /// </summary>
         /// <param name="exportDataEntity">export data entity.</param>
         /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
-        [FunctionName(nameof(HandleFailureActivityAsync))]
+        [FunctionName(FunctionNames.HandleExportFailureActivity)]
         public async Task HandleFailureActivityAsync(
             [ActivityTrigger] ExportDataEntity exportDataEntity)
         {
+            if (exportDataEntity == null)
+            {
+                throw new ArgumentNullException(nameof(exportDataEntity));
+            }
+
             await this.DeleteFileAsync(exportDataEntity.FileName);
             await this.SendFailureMessageAsync(exportDataEntity.PartitionKey);
             await this.exportDataRepository.DeleteAsync(exportDataEntity);
@@ -107,10 +90,12 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Prep.Func.Export.Activities
                 return;
             }
 
-            await this.blobContainerClient.CreateIfNotExistsAsync();
-            await this.blobContainerClient
-                    .GetBlobClient(fileName)
-                    .DeleteIfExistsAsync();
+            var blobContainerClient = this.storageClientFactory.CreateBlobContainerClient();
+
+            await blobContainerClient.CreateIfNotExistsAsync();
+            await blobContainerClient
+                .GetBlobClient(fileName)
+                .DeleteIfExistsAsync();
         }
 
         private async Task SendFailureMessageAsync(string userId)
@@ -132,7 +117,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Prep.Func.Export.Activities
 
             int maxNumberOfAttempts = 10;
             await this.botAdapter.ContinueConversationAsync(
-               botAppId: this.authorAppId,
+               botId: this.authorAppId,
                reference: conversationReference,
                callback: async (turnContext, cancellationToken) =>
                {
